+++ accounts/internal/dfns/wallet.go
@@ -115,15 +115,15 @@ func (c *dfnsClient) CreateWallet(ctx context.Context, network, name string) (*W
 	return resp, nil
 }
 
-func (c *dfnsClient) GetWalletHistory(ctx context.Context, walletID, paginationToken string, limit uint) (*WalletHistory, error) {
+func (c *dfnsClient) GetWalletHistory(ctx context.Context, walletID, paginationToken string, limit uint64) (*WalletHistory, error) {
 	header := http.Header{}
 	header.Add(authDfnsHeader, dfnsAuthHeader(ctx))
 	resp, err := dfnsCall[struct {
 		PaginationToken string `form:"paginationToken"`
-		Limit           uint   `form:"limit"`
+		Limit           uint64 `form:"limit"`
 	}, WalletHistory](ctx, c, &struct {
 		PaginationToken string `form:"paginationToken"`
-		Limit           uint   `form:"limit"`
+		Limit           uint64 `form:"limit"`
 	}{
 		PaginationToken: paginationToken,
 		Limit:           limit,
