+++ cmd/heimdall-identity-io/contract.go
@@ -160,7 +160,7 @@ type (
 		UserIDOrMasterKey string `uri:"userIdOrMasterKey" required:"true" swaggerignore:"true"`
 		WalletViewID      string `uri:"walletViewId" required:"true" swaggerignore:"true"`
 		PaginationToken   string `query:"paginationToken" form:"paginationToken" required:"false"`
-		Limit             uint   `query:"limit" form:"limit" required:"false"`
+		Limit             uint64 `query:"limit" form:"limit" required:"false"`
 	}
 	WalletViewReference struct {
 		UserIDOrMasterKey string `uri:"userIdOrMasterKey" required:"true" swaggerignore:"true"`
@@ -223,7 +223,7 @@ type (
 	GetNFTsReq  struct {
 		WalletID        string `uri:"walletId"`
 		PaginationToken string `query:"paginationToken" form:"paginationToken" required:"false"`
-		Limit           uint   `query:"limit" form:"limit" required:"false"`
+		Limit           uint64 `query:"limit" form:"limit" required:"false"`
 	}
 	NFTCollection struct {
 		WalletID        string       `json:"walletId"`
@@ -235,7 +235,7 @@ type (
 		WalletID        string `uri:"walletId"`
 		Authorization   string `header:"Authorization" swaggerignore:"true"`
 		PaginationToken string `query:"paginationToken" form:"paginationToken" required:"false"`
-		Limit           uint   `query:"limit" form:"limit" required:"false"`
+		Limit           uint64 `query:"limit" form:"limit" required:"false"`
 	}
 	WalletHistoryCollection struct {
 		WalletID        string                       `json:"walletId"`
